add an abstraction for lamdba apps
